examples/ChestMNIST_end_to_end_study/requirements.txt
torch>=1.9.0  
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.3.0
pandas>=1.3.0
scikit-learn>=1.0.0
transformers>=4.10.0
tqdm>=4.62.0
PyYAML>=5.4.0
medmnist>=2.1.0